Fix Double Validation
So as it turns out, casting `as? Int` or `as? Double` are still fundamentally flawed in Swift and do not quite work perfectly. So, to solve this, the original developer added a part to the code where it would first be converted to a string, then the number would be parsed out of that. Aside from the fact that we should not have to do this (and we unfortuanately need to).

This mehtod of validation was likely originally intentional, though no longer needed for the purpose of this validator. (i.e. Strings like "2500 hours und 14109 Berlin" should not validate to a double, because NLG Entity Extraction should be used to solve this instead). If we are trying to use this data, this should not validate, because we should also be trying just "2500" as an by itself thanks to NLG.

This commit resolves these issues by checking the converted strings for Int and Double conditions before validating from a string. It also updates the related tests so they don't fail.